+++ lib/generate_metadata.py
