+++ lib/generate_section.py
